Do not fetch files while importing collections
buildResourceJson is used for both collections and items, and always
calls buildMediaJson, which will then make API requests to the files
endpoint.
Those are useless as no media can be added to an item set.
Similarly, buildHtmlMediaJson is useless in the context of item sets.

This patch ensure that those methods are called only for items.